+++ src/com/madbros/tileminer/Menus/OptionsMenu.java
@@ -92,9 +92,9 @@ public class OptionsMenu extends Menu {
 		menuButtons = new UIButton[functions.length];
 		for(int i = 0; i < functions.length; i++) {
 			if(i == 0) {
-				menuButtons[i] = new InGameMenuUIButton(r[i].x,r[i].y, r[i].w, r[i].h,Sprites.LEFT_ARROW, functions[i], Game.batch, r[i].x+3, r[i].y+3, 0);
+				menuButtons[i] = new InGameMenuUIButton((int)r[i].x,(int)r[i].y, (int)r[i].w, (int)r[i].h,Sprites.LEFT_ARROW, functions[i], Game.batch, (int)r[i].x+3,(int) r[i].y+3, 0);
 			} else if(i == 1) {
-				menuButtons[i] = new InGameMenuUIButton(r[i].x,r[i].y, r[i].w, r[i].h,Sprites.RIGHT_ARROW, functions[i], Game.batch, r[i].x+4, r[i].y+3, 0);
+				menuButtons[i] = new InGameMenuUIButton((int)r[i].x,(int)r[i].y, (int)r[i].w, (int)r[i].h,Sprites.RIGHT_ARROW, functions[i], Game.batch, (int)r[i].x+4,(int) r[i].y+3, 0);
 				
 			} else {
 				menuButtons[i] = new PlainUIButton(r[i].x, r[i].y, r[i].w, r[i].h, strings[i], functions[i], Game.batch);
