+++ src/com/madbros/tileminer/Utils/Rect.java
@@ -5,12 +5,12 @@ import static com.madbros.tileminer.Constants.*;
 import com.madbros.tileminer.GameObjects.Actor;
 
 public class Rect {
-	public int x;
-	public int y;
-	public int w;
-	public int h;
+	public double x;
+	public double y;
+	public double w;
+	public double h;
 	
-	public Rect(int x, int y, int w, int h) {
+	public Rect(double x, double y, double w, double h) {
 		this.x = x;
 		this.y = y;
 		this.w = w;
@@ -31,21 +31,21 @@ public class Rect {
 		this(c.sRect.x - (c.absRect.x - aRect.x), c.sRect.y - (c.absRect.y - aRect.y), aRect.w, aRect.h);
 	}
 	
-	public int x1() { return x; }
+	public double x1() { return x; }
 	
-	public int y1() { return y; }
+	public double y1() { return y; }
 	
-	public int x2() { return x + w; }
+	public double x2() { return x + w; }
 	
-	public int y2() { return y + h; }
+	public double y2() { return y + h; }
 	
-	public int w() { return w; }
+	public double w() { return w; }
 	
-	public int h() { return h; }
+	public double h() { return h; }
 	
-	public int midX() { return x + w/2; }
+	public double midX() { return x + w/2; }
 	
-	public int midY() { return y + h/2; }
+	public double midY() { return y + h/2; }
 	
 	public boolean detectCollision(Rect r) {
 		if(!(x1() >= r.x2() || x2() <= r.x1() || y2() <= r.y1() || y1() >= r.y2())) {
@@ -55,19 +55,19 @@ public class Rect {
 		}
 	}
 	
-	public int getBottomCollisionDiff(Rect r) {
+	public double getBottomCollisionDiff(Rect r) {
 		return y2() - r.y1();
 	}
 	
-	public int getTopCollisionDiff(Rect r) {
+	public double getTopCollisionDiff(Rect r) {
 		return y1() - r.y2();
 	}
 	
-	public int getLeftCollisionDiff(Rect r) {
+	public double getLeftCollisionDiff(Rect r) {
 		return x1() - r.x2() ;
 	}
 	
-	public int getRightCollisionDiff(Rect r) {
+	public double getRightCollisionDiff(Rect r) {
 		return x2() - r.x1();
 	}
 }
